+++ src/app/data_harvester/disks/windows.rs
@@ -2,8 +2,17 @@
 
 use sysinfo::{DiskExt, System, SystemExt};
 
-use crate::app::filter::Filter;
-use crate::data_harvester::disks::DiskHarvest;
+use crate::{
+    app::{
+        data_harvester::disks::{DiskHarvest, IoHarvest},
+        filter::Filter,
+    },
+    utils::error,
+};
+
+pub fn get_io_usage() -> error::Result<IoHarvest> {
+    Ok(IoHarvest::default())
+}
 
 pub(crate) fn get_disk_usage(
     sys: &System, disk_filter: &Option<Filter>, mount_filter: &Option<Filter>,
